+++ src/puppet-padchat/pure-function-helpers/compatible-wei-bug.ts
